% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runbestsdp.R
\name{runbestsdp}
\alias{runbestsdp}
\title{Run the Burden ESTimate in Settlements of Displaced Populations (BESTSDP) Shiny Web Application}
\usage{
runbestsdp(launch.browser = TRUE, ...)
}
\arguments{
\item{launch.browser}{whether if you want to launch the app in an external browser.}

\item{...}{other parameters passed to shiny::runApp.}
}
\description{
Function \code{runbestsdp} is used start the BESTSDP Shiny Web Application.\cr
BestSDP is a web application created to serve as a graphical user interface for generating figures and tables for burden of conmmon communicable diseases estimate in SDP.
It was created using Shiny, a web application framework for R. \cr
}
\examples{
\donttest{
library("bestSDP")
runbestsdp(launch.browser = TRUE)
}

}
\author{
You Li,  \email{You.Li2@ed.ac.uk}
}
